+++ tests/test_di_container.py
@@ -155,6 +155,15 @@ class _FakeMemoryRuntime:
     def get_decision_window(self, decision_window_id: str, **_: object):
         return None
 
+    def upsert_communication_policy(self, **_: object):
+        raise AssertionError("not expected in this test")
+
+    def list_communication_policies(self, **_: object):
+        return []
+
+    def get_communication_policy(self, policy_id: str, **_: object):
+        return None
+
     def upsert_authority_binding(self, **_: object):
         raise AssertionError("not expected in this test")